STRIPES-950: Replace folio-org/checkout@v2 with actions/checkout@v4 in .github workflow-templates
https://folio-org.atlassian.net/browse/STRIPES-950

See discussion on #stripes-architecture on January 2024.

The folio-org fork of actions/checkout is useless, uses a version that has reached its end-of-life and should be archived.

Before that we need to replace any use of folio-org/checkout.

This ticket is to replace folio-org/checkout@v2 with actions/checkout@v4 in .github workflow-templates.